requirements.txt
streamlit
torch
torchvision
scikit-learn
matplotlib
seaborn
pandas
numpy
Pillow
tqdm